+++ examples/frame_big.c
@@ -37,7 +37,7 @@ int main(void) {
 
   static int32_t data[CHUNKSIZE];
   static int32_t data_dest[CHUNKSIZE];
-  size_t isize = CHUNKSIZE * sizeof(int32_t);
+  int32_t isize = CHUNKSIZE * sizeof(int32_t);
   blosc_timestamp_t last, current;
 
   printf("Blosc version info: %s (%s)\n",
@@ -63,7 +63,7 @@ int main(void) {
     for (int i = 0; i < CHUNKSIZE; i++) {
       data[i] = i * nchunk;
     }
-    int nchunks = blosc2_schunk_append_buffer(schunk, data, isize);
+    int64_t nchunks = blosc2_schunk_append_buffer(schunk, data, isize);
     assert(nchunks == nchunk + 1);
   }
   /* Gather some info */
@@ -72,9 +72,9 @@ int main(void) {
   blosc_set_timestamp(&current);
   double ttotal = blosc_elapsed_secs(last, current);
   printf("Compression ratio: %.1f MB -> %.1f MB (%.1fx)\n",
-         nbytes / MB, cbytes / MB, (1. * nbytes) / cbytes);
+         (double)nbytes / MB, (double)cbytes / MB, (1. * (double)nbytes) / (double)cbytes);
   printf("Time for append data to a schunk backed by a fileframe: %.3g s, %.1f MB/s\n",
-         ttotal, nbytes / (ttotal * MB));
+         ttotal, (double)nbytes / (ttotal * MB));
 
   /* Retrieve and decompress the chunks from the super-chunks and compare values */
   for (int nchunk = 0; nchunk < NCHUNKS; nchunk++) {
