+++ tests/test_bitshuffle_leftovers.c
@@ -100,7 +100,7 @@ int main(void) {
     ((uint32_t*)data)[i] = i;
   }
   /* leftovers */
-  for (int i = size / sizeof(int32_t) * sizeof(int32_t); i < size; i++) {
+  for (int i = size / (int)sizeof(int32_t) * (int)sizeof(int32_t); i < size; i++) {
     ((uint8_t*)data)[i] = (uint8_t) i;
   }
